pyproject.toml
[project]
name = "songs-dl"
version = "0.2.2"
description = "Download songs on YouTube and add metadata."
readme = "README.md"
authors = [{ name = "Laurent FAVOLE", email = "laurentfavole03@gmail.com" }]
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "License :: Public Domain",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Sound/Audio",
]
keywords = ["youtube", "music"]
dependencies = [
    "mutagen",
    "Pillow",
	"requests",
    "rich",
    "tqdm",
    "Unidecode",
    "yt-dlp",
]
requires-python = ">=3.10"

[project.optional-dependencies]
build = ["pyinstaller"]
dev = ["bump-my-version", "ruff"]
docs = ["markdown-include", "mkdocs", "mkdocs-material", "mkdocs-minify-plugin"]
lyrics = ["fpdf2"]

[project.urls]
Homepage = "https://github.com/lfavole/songs-dl"

[project.scripts]
songs-dl = "songs_dl.__main__:main"

[tool.bumpversion]
current_version = "0.2.2"
message = "chore: Release {new_version}"
commit = true
search = "{current_version}"
replace = "{new_version}"
regex = false
sign_tags = true
tag = true
tag_name = "{new_version}"
tag_message = "{new_version}"

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = "version = \"{current_version}\""
replace = "version = \"{new_version}\""

[[tool.bumpversion.files]]
filename = "songs_dl/__init__.py"
search = "__version__ = \"{current_version}\""
replace = "__version__ = \"{new_version}\""

[tool.ruff]
line-length = 120

[tool.ruff.lint]
preview = true
select = ["ALL"]
ignore = ["ANN002", "ANN003", "COM812", "CPY", "D203", "D212", "DOC201", "DOC402", "DTZ", "EXE", "FBT", "FIX", "S310", "T", "TD002", "TD003"]

[tool.ruff.format]
preview = true

[tool.setuptools]
license-files = []
packages = ["songs_dl"]
